Update with Factoring: - move `backend/db/models` to `backend/models` - move `db/mongoose.js` to services - refactor `authenticate` from `index.js` as on middleware - adding config file `config/default.js` to store jwtPrivateKey - remove jwtPrivateKey getter in `User`
