+++ backend/config/default.json
@@ -1,0 +1,4 @@
+{
+    "jwtPrivateKey": "51778657246321226641fsdklafjasdkljfsklfjd7148924065"
+}
+
