--- backend/db/models/index.js
@@ -1,5 +1,0 @@
-const { User } = require('./user.model');
-
-module.exports = {
-    User
-}
