+++ backend/index.js
@@ -6,15 +6,15 @@ const util = require('util');
 const jwt = require('jsonwebtoken');
 const bodyParser = require('body-parser');
 
+const { User } = require('./models');
+const { auth } = require('./middleware/auth');
+const pollRouter = require('./routes/poll');
+
 const express = require('express');
 const app = express();
 
-// Load in the mongoose models
-const { mongoose } = require('./db/mongoose');
-const { User } = require('./db/models');
-
-// Router
-const pollRouter = require('./routes/poll');
+// Setup mongoose
+require("./startup/mongoose")();
 
 // Blockchain Network
 let network = require('./fabric/network.js');
@@ -24,7 +24,6 @@ const config = JSON.parse(configJSON);
 const appAdmin = config.appAdmin;
 
 /* MIDDLEWARE  */
-// Load middleware
 app.use(bodyParser.json());
 
 // CORS HEADERS MIDDLEWARE
@@ -38,24 +37,6 @@ app.use(function (req, res, next) {
     );
     next();
 });
-
-// check whether the request has a valid JWT access token
-let authenticate = (req, res, next) => {
-    let token = req.header('x-access-token');
-
-    // verify the JWT
-    jwt.verify(token, User.getJWTSecret(), (err, decoded) => {
-        if (err) {
-            // there was an error
-            // jwt is invalid - * DO NOT AUTHENTICATE *
-            res.status(401).send(err);
-        } else {
-            // jwt is valid
-            req.user_id = decoded._id;
-            next();
-        }
-    });
-}
 
 // Verify Refresh Token Middleware (which will be verifying the session)
 let verifySession = (req, res, next) => {
@@ -77,7 +58,7 @@ let verifySession = (req, res, next) => {
         // if the code reaches here - the user was found
         // therefore the refresh token exists in the database - but we still have to check if it has expired or not
 
-        req.user_id = user._id;
+        req.userID = user._id;
         req.userObject = user;
         req.refreshToken = refreshToken;
 
@@ -110,7 +91,7 @@ let verifySession = (req, res, next) => {
 /* END MIDDLEWARE  */
 
 /* ROUTE HANDLERS */
-app.use('/poll', pollRouter);
+app.use('/poll',auth, pollRouter);
 
 /* USER ROUTES */
 
@@ -168,7 +149,6 @@ app.post('/users/login', (req, res) => {
             });
         }).then(async (authTokens) => {
             // Now we construct and send the response to the user with their auth tokens in the header and the user object in the body
-            
             // This is how to connect to Blockchain network with userId
             let connection = await network.connectToNetwork(user._id.toString());
             
@@ -188,7 +168,7 @@ app.post('/users/login', (req, res) => {
  * Purpose: generates and returns an access token
  */
 app.get('/users/me/access-token', verifySession, (req, res) => {
-    // we know that the user/caller is authenticated and we have the user_id and user object available to us
+    // we know that the user/caller is authenticated and we have the userID and user object available to us
     req.userObject.generateAccessAuthToken().then((accessToken) => {
         res.header('x-access-token', accessToken).send({ accessToken });
     }).catch((e) => {
