+++ backend/models/index.js
@@ -1,0 +1,8 @@
+const { User } = require('./user');
+const { Poll, validatePoll } = require('./poll');
+
+module.exports = {
+    User,
+    Poll,
+    validatePoll
+}
