+++ backend/package.json
@@ -14,15 +14,16 @@
     "@sendgrid/mail": "^7.4.2",
     "bcryptjs": "^2.4.3",
     "body-parser": "^1.19.0",
+    "config": "^3.3.6",
+    "eslint": "^6.0.1",
     "express": "^4.17.1",
+    "fabric-ca-client": "~1.4.2",
+    "fabric-client": "^1.4.10",
+    "fabric-network": "~1.4.2",
     "joi": "^17.4.0",
     "jsonwebtoken": "^8.5.0",
     "lodash": "^4.17.11",
     "mongoose": "^5.4.11",
-    "nodemon": "^1.19.1",
-    "eslint": "^6.0.1",
-    "fabric-ca-client": "~1.4.2",
-    "fabric-client": "^1.4.10",
-    "fabric-network": "~1.4.2"
+    "nodemon": "^1.19.1"
   }
 }
