+++ backend/routes/poll.js
@@ -1,6 +1,6 @@
 const express = require("express");
 const router = express.Router();
-const { Poll , validatePoll} = require("../models/poll");
+const { Poll, validatePoll } = require('../models');
 
 router.get("/", async (req, res) => {
     res.send("hello front poll router");
@@ -17,22 +17,23 @@ router.get("/:id", async (req, res) => {
 
 router.post("/", async (req, res) => {
 
-    const { error } = validatePoll(req.body);
-    if (error) return res.status(400).send(error.details[0].message);
+  const { error } = validatePoll(req.body);
+  if (error) return res.status(400).send(error.details[0].message);
 
-    const poll = new Poll({
-      title: req.body.title,
-      type: req.body.type,
-      deadline: req.body.deadline,
-      accessLevel: req.body.accessLevel,
-      isAnonymousModeOn: req.body.isAnonymousModeOn,
-      isHiddenUntilDeadline: req.body.isHiddenUntilDeadline,
-      canVotersSeeResults: req.body.canVotersSeeResults,
-      questions: req.body.questions
-    });
+  const poll = new Poll({
+    title: req.body.title,
+    type: req.body.type,
+    host: req.userID,
+    deadline: req.body.deadline,
+    accessLevel: req.body.accessLevel,
+    isAnonymousModeOn: req.body.isAnonymousModeOn,
+    isHiddenUntilDeadline: req.body.isHiddenUntilDeadline,
+    canVotersSeeResults: req.body.canVotersSeeResults,
+    questions: req.body.questions
+  });
 
-    await poll.save();
-    res.send(poll);
+  await poll.save();
+  res.send(poll);
 });
 
 module.exports = router;
